requirements.txt
# Warning: this file only contains development requirements!
# In reality this app should use the open edX dependencies that already are installed.
celery==3.1.26.post2
django-celery==3.3.1
edx-celeryutils==0.5.0
requests==2.23.0
-e git+git://github.com/edx/edx-platform.git@open-release/juniper.master#egg=XModule&subdirectory=common/lib/xmodule
-e git+git://github.com/edx/edx-platform.git@open-release/juniper.master#egg=Open-edX